samples/python/requirements.txt
fastapi
uvicorn
azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi
